Changed HP, DEF, ATK, SCORE from ints to doubles

+++ BaseCharacter.cc
@@ -2,14 +2,14 @@
 #include "Character.h"
 #include "Observer.h"
 
-BaseCharacter::BaseCharacter(size_t x, size_t y, char icon, int hp, int maxHp, int atk, int def) : x{x}, y{y}, icon{icon}, hp{hp}, maxHp{maxHp}, atk{atk}, def{def} {}
+BaseCharacter::BaseCharacter(size_t x, size_t y, char icon, double hp, double maxHp, double atk, double def) : x{x}, y{y}, icon{icon}, hp{hp}, maxHp{maxHp}, atk{atk}, def{def} {}
 
 void BaseCharacter::hit(Character &other) {
 	// Class specific attacks can override this. By default, deal the attack
 	other.damage(*this, atk);
 }
 
-void BaseCharacter::damage(Character &other, int damage) {
+void BaseCharacter::damage(Character &other, double damage) {
 	(void) other;
 	// Class specific defences can override this. By default, take the damage
 	hp -= damage;
@@ -22,11 +22,11 @@ void BaseCharacter::deathCheck() {
 	}
 }
 
-int BaseCharacter::score() { return gold; }
-int BaseCharacter::getHP() { return hp; }
-int BaseCharacter::getMaxHP() { return maxHp; }
-int BaseCharacter::getAtk() { return atk; }
-int BaseCharacter::getDef() { return def; }
+double BaseCharacter::score() { return gold; }
+double BaseCharacter::getHP() { return hp; }
+double BaseCharacter::getMaxHP() { return maxHp; }
+double BaseCharacter::getAtk() { return atk; }
+double BaseCharacter::getDef() { return def; }
 void BaseCharacter::addGold(int value) { gold += value; }
 int BaseCharacter::getGold() { return gold; }
 int BaseCharacter::getFloor() { return floor; }
